Cherry-pick fd588f939cf456f39326 fix broken encrypted root mount.
cryptsetup: Use openssl instead of libgcrypt

When libgcrypt added support for libcap, it started dropping
capabilities of the cryptsetup process. This meant that cryptsetup was
unable to communicate with device mapper. This patch fixes that issue by
removing the dependency on libgcrypt and instead uses openssl for the
implementation of low-level cryptographic functions.

Fixes #7859.